Moved LogsTrait one level up So it can be used in unit tests.
