+++ Tests/Unit/src/DBUtilities/TestCase.php
@@ -7,20 +7,23 @@
 
 namespace Screenfeed\AutoWPDB\Tests\Unit\src\DBUtilities;
 
-use Screenfeed\AutoWPDB\Tests\Unit\TestCase as BaseTestCase;
 use Screenfeed\AutoWPDB\Tests\Fixtures\src\DBUtilitiesUnit;
+use Screenfeed\AutoWPDB\Tests\LogsTrait;
+use Screenfeed\AutoWPDB\Tests\Unit\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase {
+	use LogsTrait;
+
 	protected $table_name        = 'wp_foobar';
 	protected $target_table_name = 'wp_targettable';
-	public $logs;
 
 	/**
 	 * Prepares the test environment before each test.
 	 */
 	public function setUp(): void {
 		parent::setUp();
-		$this->logs = [];
+
+		$this->empty_logs();
 		DBUtilitiesUnit::$mocks = [];
 	}
 
@@ -29,11 +32,8 @@ abstract class TestCase extends BaseTestCase {
 	 */
 	public function tearDown(): void {
 		parent::tearDown();
-		$this->logs = [];
-		DBUtilitiesUnit::$mocks = [];
-	}
 
-	public function log( $message ) {
-		$this->logs[] = $message;
+		$this->empty_logs();
+		DBUtilitiesUnit::$mocks = [];
 	}
 }
