+++ Tests/Unit/src/DBUtilities/createTable.php
@@ -32,7 +32,7 @@ class Test_CreateTable extends TestCase {
 		);
 
 		$this->assertTrue( $result );
-		$this->assertCount( 0, $this->logs );
+		$this->assertCount( 0, $this->get_logs() );
 	}
 
 	public function testShouldReturnFalseWhenDBError() {
@@ -56,8 +56,8 @@ class Test_CreateTable extends TestCase {
 		$error = sprintf( 'Error while creating the DB table %s: %s', $this->table_name, $db_error );
 
 		$this->assertFalse( $result );
-		$this->assertContains( $error, $this->logs );
-		$this->assertCount( 1, $this->logs );
+		$this->assertContains( $error, $this->get_logs() );
+		$this->assertCount( 1, $this->get_logs() );
 	}
 
 	public function testShouldReturnFalseWhenTableError() {
@@ -79,8 +79,8 @@ class Test_CreateTable extends TestCase {
 		$error = sprintf( 'Creation of the DB table %s failed.', $this->table_name );
 
 		$this->assertFalse( $result );
-		$this->assertContains( $error, $this->logs );
-		$this->assertCount( 1, $this->logs );
+		$this->assertContains( $error, $this->get_logs() );
+		$this->assertCount( 1, $this->get_logs() );
 	}
 
 	public function testShouldFailWithoutLogging() {
@@ -102,7 +102,7 @@ class Test_CreateTable extends TestCase {
 		);
 
 		$this->assertFalse( $result );
-		$this->assertCount( 0, $this->logs );
+		$this->assertCount( 0, $this->get_logs() );
 	}
 
 	public function createMocks( $db_error = '' ) {
